create a SeedData class and update double to decimal

+++ BreweryWholesale/Models/BeerDto.cs
@@ -6,8 +6,8 @@ namespace BreweryWholesale.Models
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public double AlcoholContent { get; set; }
-        public double Price { get; set; }
+        public decimal AlcoholContent { get; set; }
+        public decimal Price { get; set; }
 
         public static IEnumerable<BeerDto> BeerDtos(IEnumerable<Beer> beers)
         {
